requirements.txt
twilio==8.10.0
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pillow==10.1.0
python-telegram-bot==20.7 